+++ game/models.py
@@ -1,19 +1,22 @@
 import random
 from game.language.letters_spa import Letters
+
 class Tile:
     def __init__(self, letter, value):
         self.letter = letter
         self.value = value
     def __repr__(self):
         return f"{self.letter}:{self.value}"
+
 class BagTiles:
     def __init__(self):
         self.tiles = []
+        self.initBagTiles()
     
     def createBagTiles(self, dict, n):
         for key, value in dict.items():
-                for _ in range(0, value):
-                    self.tiles.append(Tile(key, n))
+            for _ in range(0, value):
+                self.tiles.append(Tile(key, n))
 
     def initBagTiles(self):
         unpacking = [(Letters.onePoint, 1),
@@ -26,16 +29,8 @@ class BagTiles:
         for z in unpacking:
             self.createBagTiles(*z)
         return self.tiles;
-    
-    def getTiles(self):
-        bag = self.initBagTiles()
-        random.shuffle(bag)
-        for _ in range(0, 91):
-            self.tiles.pop()
-        return self.tiles
           
     def take(self, count):
-        self.getTiles()
         for _ in range(count):
             self.tiles.pop()
         return self.tiles