+++ tests/test_models.py
@@ -1,8 +1,6 @@
 import unittest
-from game.models import (
-    BagTiles,
-    Tile,
-)
+from game.models import (BagTiles, Tile)
+
 global bag 
 bag = BagTiles()
 from unittest.mock import patch
@@ -17,30 +15,19 @@ class TestBagTiles(unittest.TestCase):
     @patch('random.shuffle')
     def test_bag_tiles(self, patch_shuffle):
         s = BagTiles()
-        s.getTiles()
-        self.assertEqual(
-            len(s.tiles),
-            7,
-        )
+        self.assertEqual(len(s.tiles), 98)
 
     def test_take(self):
         bag = BagTiles()
         bag.take(2)
-        self.assertEqual(
-            len(bag.tiles),
-            5,
-        )
+        self.assertEqual(len(bag.tiles), 96)
 
     def test_put(self):
         bag = BagTiles()
+        bag.take(7)
         put_tiles = [Tile('Z', 1), Tile('Y', 1)]
         bag.put(put_tiles)
-        self.assertEqual(
-            len(bag.tiles),
-            2,
-        )
+        self.assertEqual(len(bag.tiles), 93)
 
 if __name__ == '__main__':
     unittest.main()
-
-print(bag.tiles)